# packages/regpoly-web/tests/fixtures/openapi.v2.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/primitive-searches/{run_id}/duplicate": {
      "get": {
        "operationId": "v2_primitive_duplicate_api_v2_primitive_searches__run_id__duplicate_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_PrefillResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V2 Primitive Duplicate",
        "tags": [
          "v2"
        ]
      }
    },
    "/api/v2/primitive-searches/{run_id}/history": {
      "get": {
        "operationId": "v2_primitive_history_api_v2_primitive_searches__run_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "points",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 1000,
              "minimum": 1,
              "title": "Points",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V2 Primitive History",
        "tags": [
          "v2"
        ]
      }
    },
...
    "/api/v2/tempering-searches/{run_id}/duplicate": {
      "get": {
        "operationId": "v2_tempering_duplicate_api_v2_tempering_searches__run_id__duplicate_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_TemperingPrefill"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V2 Tempering Duplicate",
        "tags": [
          "v2"
        ]
      }
    },
    "/api/v2/tempering-searches/{run_id}/history": {
      "get": {
        "operationId": "v2_tempering_history_api_v2_tempering_searches__run_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "points",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 1000,
              "minimum": 1,
              "title": "Points",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V2 Tempering History",
        "tags": [
          "v2"
        ]
      }
    },
...
  }
}